+++ build.sbt
@@ -548,7 +548,7 @@ lazy val testkitCore = project
   .settings(CommonSettings.prodSettings: _*)
   .settings(name := "bitcoin-s-testkit-core",
             libraryDependencies ++= Deps.testkitCore)
-  .dependsOn(core, crypto)
+  .dependsOn(asyncUtils, core, crypto)
 
 lazy val testkit = project
   .in(file("testkit"))
